big refactor, start trying to make a pam module
This refactors damn near everything to use separate crates, including
putting the main auth logic into a crate so that it can be called
from `tp-exec` (a pam_exec shim that we can ship for platforms we
can't build dynamic binaries for) and pam itself. This also adds the
boilerplate for a proper pam module.

The big annoying part about this pam module is that they don't seem to
have the pam_getenvlist function implemented for "dear god this is
totally unsafe" kinds of reasons. However I think I have a way to at
least mitigate a lot of the danger.

Signed-off-by: Christine Dodrill <xe@tailscale.com>