cmd/pam_tailscale: support ngircd authentication
ngircd sets the reverse DNS hostname of the connecting client as the
PAM_RHOST instead of the IP address like sshd does. This change makes
pam_tailscale perform a DNS lookup on the supplied PAM_RHOST instead of
just assuming it's an IP address. If it finds a result, then the rhost
is set to that result. If not, then autentication fails with
PAM_AUTHTOK_ERR and a summary of what was going on is printed to syslog.

There is probably a more elegant way to do this, but this works for now.

Signed-off-by: Xe <me@christine.website>